Add files via upload
This project aims at taking a step into the field of data security by creating a machine that is capable of data encryption. The input is taken from the user at runtime in the form   of a text which is stored in an array. This string array is then passed through the machine which encrypts the data and this encrypted data is then transmitted as a ciphertext. At the receiving end, the data can only be decrypted if the receiving machine has the correct encryption key. If the encryption key matches, the ciphertext undergoes decryption, and the text sent by the sender is displayed at the receiver’s end.